rm docker add hub binaries

+++ scripts/storehook.js
@@ -22,47 +22,45 @@
 * <http://resources.spinalcom.com/licenses.pdf>.
 */
 
-const fs = require('fs');
-const { exec } = require('child_process');
-const path = require('path');
+const fs = require("fs");
+const { exec } = require("child_process");
+const path = require("path");
 
-var hookDir = path.resolve('../.hooks/');
+var hookDir = path.resolve("../.hooks/");
 
 var hookPath = {
-  origin: path.resolve('./hooks/postinstall'),
-  target: path.resolve('../.hooks/postinstall')
-}
+  origin: path.resolve("./hooks/postinstall"),
+  target: path.resolve("../.hooks/postinstall")
+};
 
 var hookScriptPath = {
-  origin: path.resolve('./hooks/run_post.js'),
-  target: path.resolve('../.hooks/run_post.js')
-}
+  origin: path.resolve("./hooks/run_post.js"),
+  target: path.resolve("../.hooks/run_post.js")
+};
 
-if (!fs.existsSync(hookPath.target)) { 
+if (!fs.existsSync(hookPath.target)) {
   setupHook();
 }
 
 function setupHook() {
-  exec('ls', (err, stdout, stderr) => {
+  exec("ls", err => {
     if (err) {
       console.error(`exec error: ${err}`);
       return;
     }
   });
 
-  if (!fs.existsSync(hookDir)) { 
+  if (!fs.existsSync(hookDir)) {
     fs.mkdir(hookDir);
   }
-
-  let content = 'node ' + hookScriptPath.target;
 
   // TODO: hacer esto sync
 
   copyRecursiveSync(hookScriptPath.origin, hookScriptPath.target);
   copyRecursiveSync(hookPath.origin, hookPath.target);
-  fs.chmodSync(hookPath.target, '777');
-  console.log('New Spinal module installed.');
-/*
+  fs.chmodSync(hookPath.target, "777");
+  console.log("New Spinal module installed.");
+  /*
   fs.writeFile(hookPath.target, content, { flag : 'w' }, function (err) {
     if (err) return console.log(err);
 
@@ -78,16 +76,17 @@ function setupHook() {
 function copyRecursiveSync(src, dest) {
   var exists = fs.existsSync(src);
   var stats = exists && fs.statSync(src);
-  if (exists && fs.existsSync(dest))
-      fs.unlinkSync(dest);
+  if (exists && fs.existsSync(dest)) fs.unlinkSync(dest);
   var isDirectory = exists && stats.isDirectory();
   if (exists && isDirectory) {
     fs.mkdirSync(dest);
     fs.readdirSync(src).forEach(function(childItemName) {
-      copyRecursiveSync(path.join(src, childItemName),
-                        path.join(dest, childItemName));
+      copyRecursiveSync(
+        path.join(src, childItemName),
+        path.join(dest, childItemName)
+      );
     });
   } else {
     fs.linkSync(src, dest);
   }
-};
+}
